refactor: move session policy helpers under session lib
